Canvas initialization drawing
Canvas is now drawing with correct perspective-- now realizing that I'll need to constantly be updating the state of each canvas element and redrawing it as the user interacts with it. Thank god I set this up using objects. This is something of a larger solution and will take more time. Will check in with Kim about the timetable she's expecting for this delivery :)